Add `PrunedStore` and use it in Wasmi's executor (#1449)
* add TypedStore internal parts of Store

* put TypedStore's data into a Box

This is important to enforce that all TypedStore<T> are of equal sizes.

* rename TypedStore to TypedStoreInner

* add test asserting equal sizes for all TypedStoreInner<T>

* move data field to last position

* add PrunedStore definitions and tests

* change expect -> allow

* rename store field to pruned

* move id field from PrunedStore to Store<T>

This way PrunedStore is pointer sized.

* make Wasmi executor use Store::call_host_func

This is an API can can more easily be wrapped for an indirect call via a PrunedStore.

* rename FuncParams to FuncInOut

This is a better name since this type manages both, function parameters and results.

* move func_inout.rs into wasmi::func submodule

* fix intra doc links

* add call_host_func trampoline to Store<T>

This allows to call host functions via a PrunedStore.

* improve Debug impl for ResourceLimiterQuery

* apply rustfmt

* properly silence warning

* use proper signatures in PrunedStore::inner[_mut] methods

This way they cannot be used with foo.bar() syntax to disambiguate them with their Deref[Mut] impls.

* make PrunedStore::restore a static method

* add PrunedStore::call_host_func method

* make PrunedStore a transparent type

It no longer holds a reference to a Store<Prune> but simply is the shape of a Store<Prune> and now has #[repr(transparent)] to make transmutation sound.

* generalize PrunedStore restoration

* refactor and clean-up new design

* make use of PrunedStore in call

This is just for demo purposes (it works) and not final code.
Call hooks from Wasm -> Host are now handled at the TypedStore.

* add TypedStore::store_inner_and_resource_limiter_ref

Also make RestorePrunedWrapper::restore infallible.

* port Wasmi's executor to PrunedStore

* apply clippy suggestions

* clean-up PruneStore::inner[_mut] methods

* make PrunedStore::restore method private

* put #[inline] on PrunedStore::restore method

* add unlikely to PrunedStore::restore check

* unsilence warnings

* put some inlines where it makes sense

* make Store::prune method crate private

* revert some unnecessary changes compared to `main`

* use imported type_name

* fix docs

* deduplicate code in PrunedStore

* add missing docs to PrunedStore API

* apply rustfmt

* apply inline(always) to execute_branch_binop op-handlers

* fix most of the call instruction regressions

* remove commented out inline annotations

* put inline on merge_call_frame

This prevents LLVM from generating very weird phi gigantic nodes.